+++ src/infrastructure/providers/dotnet/test/dotnetXmlParserFactory/createDependenciesFromXml.tests.ts
@@ -6,23 +6,17 @@ const assert = require('assert');
 
 export default {
 
-  "returns empty when no matches found": () => {
-    const includeNames = []
-    const results = createDependenciesFromXml('', includeNames);
-    assert.equal(results.length, 0);
-  },
-
-  "returns empty when no dependency entry names match": () => {
-    const includeNames = ["non-dependencies"]
-    const results = createDependenciesFromXml(
-      Fixtures.createDependenciesFromXml.test,
-      includeNames
-    );
-    assert.equal(results.length, 0);
-  },
+  "returns empty when no matches found": [
+    [[], ''],
+    [["non-dependencies"], Fixtures.createDependenciesFromXml.test],
+    (testIncludeNames, expected) => {
+      const results = createDependenciesFromXml(expected, testIncludeNames);
+      assert.equal(results.length, 0);
+    }
+  ],
 
   "extracts dependency entries from dotnet xml": () => {
-    const includeNames = ["PackageReference", "PackageVersion"]
+    const includeNames = ["Sdk", "PackageReference", "PackageVersion"]
     const results = createDependenciesFromXml(
       Fixtures.createDependenciesFromXml.test,
       includeNames